requirements.txt
msal>=1.25.0
httpx>=0.25.0
python-dotenv>=1.0.0
